+++ lib/ReactViews/FeatureInfo/FeatureInfoDownload.tsx
@@ -1,6 +1,5 @@
-import { TFunction } from "i18next";
 import React from "react";
-import { withTranslation } from "react-i18next";
+import { withTranslation, TFunction } from "react-i18next";
 import DataUri from "../../Core/DataUri";
 import filterOutUndefined from "../../Core/filterOutUndefined";
 import JsonValue, { JsonObject } from "../../Core/Json";